% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improvedApriori.R
\name{binarizeData}
\alias{binarizeData}
\title{R function for binarizing safety data for implementing improvedApriori function}
\usage{
binarizeData(file = file, numdrug = 5, numae = 13, numreports = 13)
}
\description{
R function for binarizing safety data for implementing improvedApriori function
}
